Refractor note pop up so it can reusable
