+++ finance/lib/features/transactions/presentation/widgets/transaction_list.dart
@@ -15,6 +15,7 @@ import '../../../../shared/utils/responsive_layout_builder.dart';
 import '../../../../shared/utils/performance_optimization.dart';
 import '../../../../shared/utils/no_overscroll_behavior.dart';
 import '../../../../shared/widgets/animations/tappable_widget.dart';
+import '../../../../shared/widgets/dialogs/note_popup.dart';
 
 /// Widget that displays a list of transactions grouped by date
 class TransactionList extends StatelessWidget {
@@ -260,25 +261,6 @@ class TransactionTile extends StatelessWidget {
   final Transaction transaction;
   final Category? category;
 
-  // ---------------------------------------------------------------------
-  // Static state for managing a single note-popup overlay application-wide
-  // ---------------------------------------------------------------------
-  static OverlayEntry? _noteOverlayEntry;
-  static AnimationController? _noteAnimationController;
-  static VoidCallback? _scrollListener;
-  static ScrollPosition? _scrollPosition;
-
-  /// Dismisses the current popup (if any) with a fade-out animation.
-  static void _dismissCurrentPopup() {
-    if (_noteAnimationController == null) return;
-    // If already reversing, do nothing.
-    if (_noteAnimationController!.status == AnimationStatus.reverse ||
-        _noteAnimationController!.status == AnimationStatus.dismissed) {
-      return;
-    }
-    _noteAnimationController!.reverse();
-  }
-
   @override
   Widget build(BuildContext context) {
     // Phase 5: Cache theme data for performance (Phase 1 pattern)
@@ -339,7 +321,6 @@ class TransactionTile extends StatelessWidget {
             trailing: Row(
               mainAxisSize: MainAxisSize.min,
               children: [
-                // Transaction note button
                 if (transaction.note != null && transaction.note!.isNotEmpty) ...[
                   Builder(builder: (context) {
                     return GestureDetector(
@@ -348,7 +329,7 @@ class TransactionTile extends StatelessWidget {
                             context.findRenderObject() as RenderBox;
                         final position = renderBox.localToGlobal(Offset.zero);
                         final size = renderBox.size;
-                        _showNotePopup(context, transaction.note!, position, size);
+                        NotePopup.show(context, transaction.note!, position, size);
                       },
                       child: SvgPicture.asset(
                         'assets/icons/icon_note.svg',
@@ -386,125 +367,5 @@ class TransactionTile extends StatelessWidget {
         ),
       ),
     );
-  }
-
-  void _showNotePopup(
-      BuildContext context, String note, Offset position, Size size) {
-    // ------------------------------------------------------------
-    // Enhanced overlay management: single entry + scroll dismissal
-    // ------------------------------------------------------------
-
-    // 1. If there is an existing popup, dismiss it first.
-    _dismissCurrentPopup();
-
-    final overlay = Overlay.of(context);
-    if (overlay == null) return;
-
-    // 2. Calculate popup position (same logic as before)
-    final screenSize = MediaQuery.of(context).size;
-    const double popupMaxWidth = 250.0;
-    const double popupPadding = 16.0;
-    const double verticalOffset = 8.0;
-
-    double left = position.dx + (size.width / 2) - (popupMaxWidth / 2);
-    double top = position.dy + size.height + verticalOffset;
-
-    if (left + popupMaxWidth + popupPadding > screenSize.width) {
-      left = screenSize.width - popupMaxWidth - popupPadding;
-    }
-    if (left < popupPadding) left = popupPadding;
-
-    if (top + 100 > screenSize.height - popupPadding) {
-      top = position.dy - 100 - verticalOffset;
-    }
-    if (top < popupPadding) top = popupPadding;
-
-    // 3. Prepare animation controller (store in static var)
-    final controller = AnimationController(
-      duration: const Duration(milliseconds: 250),
-      reverseDuration: const Duration(milliseconds: 250),
-      vsync: Navigator.of(context),
-    );
-
-    late OverlayEntry entry;
-    entry = OverlayEntry(
-      builder: (context) => Positioned(
-        left: left,
-        top: top,
-        child: IgnorePointer(
-          ignoring: true,
-          child: FadeTransition(
-            opacity: CurvedAnimation(parent: controller, curve: Curves.easeInOut),
-            child: Material(
-              type: MaterialType.transparency,
-              child: Container(
-                constraints: const BoxConstraints(
-                  maxWidth: popupMaxWidth,
-                  minWidth: 120.0,
-                ),
-                padding: const EdgeInsets.all(12),
-                decoration: BoxDecoration(
-                  color: Theme.of(context).cardColor,
-                  borderRadius: BorderRadius.circular(12),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withOpacity(0.15),
-                      blurRadius: 20,
-                      offset: const Offset(0, 8),
-                      spreadRadius: 0,
-                    ),
-                    BoxShadow(
-                      color: Colors.black.withOpacity(0.08),
-                      blurRadius: 6,
-                      offset: const Offset(0, 2),
-                      spreadRadius: 0,
-                    ),
-                  ],
-                ),
-                child: AppText(
-                  note,
-                  maxLines: 5,
-                  overflow: TextOverflow.ellipsis,
-                  fontSize: 14,
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    // 4. Store static refs & insert overlay
-    _noteOverlayEntry = entry;
-    _noteAnimationController = controller;
-    overlay.insert(entry);
-    controller.forward();
-
-    // 5. Attach a scroll listener so scrolling dismisses the popup.
-    _scrollPosition = Scrollable.of(context)?.position;
-    if (_scrollPosition != null) {
-      _scrollListener = () => _dismissCurrentPopup();
-      _scrollPosition!.addListener(_scrollListener!);
-    }
-
-    // 6. Auto-dismiss after 3 seconds
-    Future.delayed(const Duration(seconds: 3), _dismissCurrentPopup);
-
-    // 7. Clean up when animation completes
-    controller.addStatusListener((status) {
-      if (status == AnimationStatus.dismissed) {
-        _noteOverlayEntry?.remove();
-        _noteOverlayEntry = null;
-
-        controller.dispose();
-        _noteAnimationController = null;
-
-        if (_scrollPosition != null && _scrollListener != null) {
-          _scrollPosition!.removeListener(_scrollListener!);
-        }
-        _scrollPosition = null;
-        _scrollListener = null;
-      }
-    });
   }
 }
